01-intro-to-rag/requirements.txt
pinecone>=7.0,<8.0
openai>=1.50
python-dotenv>=1.0
pandas>=2.2
pyarrow>=15.0
numpy>=1.26
matplotlib>=3.8
tqdm>=4.66
ipykernel>=6.29
requests>=2.31
beautifulsoup4>=4.12
